+++ server/routes/business.routes.js
@@ -11,4 +11,11 @@ router.post('/', businessController.createEntity);
 router.put('/:id', businessController.updateEntity);
 router.delete('/:id', businessController.deleteEntity);
 
+// File management routes
+const businessFilesController = require('../controllers/business-files.controller');
+router.get('/:id/files', businessFilesController.getEntityFiles);
+router.post('/:id/files', businessController.upload ? businessController.upload.single('file') : require('../middleware/upload').upload.single('file'), businessFilesController.uploadFile);
+router.delete('/:id/files/:fileId', businessFilesController.deleteFile);
+router.get('/:id/files/:fileId/download', businessFilesController.downloadFile);
+
 module.exports = router;